Fix imports after refactor
